requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
urllib3==1.26.2
